sections/3-build-a-retrieval-project/fastapi-backend/requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
weaviate-client
supabase
